+++ crates/utils/transaction-generator/src/generators/basic.rs
@@ -2,29 +2,29 @@ use std::sync::Arc;
 
 use derivative::Derivative;
 use serde::{Deserialize, Serialize};
+use sov_modules_api::prelude::axum::async_trait;
 use sov_modules_api::Spec;
 
 use super::bank::harness_interface::BankHarness;
-use super::bank::{BankChangeLogEntry, BankMessageGenerator, BankTag};
+use super::bank::{BankChangeLogDiscriminant, BankChangeLogEntry, BankTag};
 use super::factory::CallMessageFactory;
 use super::value_setter::{
-    ValueSetterChangeLogEntry, ValueSetterHarness, ValueSetterMessageGenerator,
+    ValueSetterChangeLogDiscriminant, ValueSetterChangeLogEntry, ValueSetterHarness,
 };
-use crate::interface::traits::CallMessageGenerator;
-use crate::HarnessModule;
+use crate::{ChangelogEntry, HarnessModule};
 
 /// A basic call message generator factory that can be used with modules internal to the sovereign sdk
 pub type BasicCallMessageFactory<S, RT, Acct = ()> =
-    CallMessageFactory<S, RT, BasicTag, BasicChangeLogEntry<S>, BasicClientConfig, Acct>;
+    CallMessageFactory<S, RT, BasicTag, BasicChangeLogEntry<S>, Acct>;
 /// A helper type that corresponds to bank modules compatible with the basic harness
 pub type BasicBankHarness<S, RT, Acct = ()> =
-    BankHarness<S, RT, BasicTag, BasicChangeLogEntry<S>, BasicClientConfig, Acct>;
+    BankHarness<S, RT, BasicTag, BasicChangeLogEntry<S>, Acct>;
 /// A helper type that corresponds to value setter modules compatible with the basic harness
 pub type BasicValueSetterHarness<S, RT, Acct = ()> =
-    ValueSetterHarness<S, RT, BasicTag, BasicChangeLogEntry<S>, BasicClientConfig, Acct>;
+    ValueSetterHarness<S, RT, BasicTag, BasicChangeLogEntry<S>, Acct>;
 /// A helper type that contains a reference to a basic module
 pub type BasicModuleRef<S, RT, Acct = ()> =
-    Arc<dyn HarnessModule<S, RT, BasicTag, BasicChangeLogEntry<S>, BasicClientConfig, Acct>>;
+    Arc<dyn HarnessModule<S, RT, BasicTag, BasicChangeLogEntry<S>, Acct>>;
 
 /// The set of tags supported by the [`BasicCallMessageFactory`].
 #[derive(Clone, Copy, Derivative, Debug, derive_more::From)]
@@ -37,7 +37,7 @@ pub enum BasicTag {
 }
 
 /// The set of change log entries supported by the [`BasicCallMessageFactory`].
-#[derive(Clone, Debug, derive_more::From, PartialEq, Deserialize, Serialize)]
+#[derive(Clone, Debug, derive_more::From, Deserialize, Serialize)]
 #[serde(rename_all = "snake_case")]
 pub enum BasicChangeLogEntry<S: Spec> {
     /// Changes from the bank module
@@ -46,39 +46,46 @@ pub enum BasicChangeLogEntry<S: Spec> {
     ValueSetter(ValueSetterChangeLogEntry),
 }
 
-/// Asserts all the [`BasicChangeLogEntry`] against the existing state
-pub async fn assert_logs_against_state<S: Spec>(
-    logs: Vec<BasicChangeLogEntry<S>>,
-    config: &BasicClientConfig,
-) -> anyhow::Result<()> {
-    let mut seen_entries: Vec<BasicChangeLogEntry<S>> = Vec::new();
+/// Helper struct that can be used to discriminate between different [`BasicChangeLogEntry`]s.
+#[derive(Debug, PartialEq, Eq, Derivative)]
+#[derivative(Hash(bound = ""))]
+pub enum BasicChangeLogDiscriminant<S: Spec> {
+    /// Discriminants from the bank module
+    Bank(BankChangeLogDiscriminant<S>),
+    /// Discriminants from the value setter module
+    ValueSetter(ValueSetterChangeLogDiscriminant),
+}
 
-    for log in logs.iter().rev() {
-        if seen_entries.contains(log) {
-            continue;
-        }
+#[async_trait]
+impl<S: Spec> ChangelogEntry for BasicChangeLogEntry<S> {
+    type ClientConfig = BasicClientConfig;
 
-        seen_entries.push(log.clone());
+    type Discriminant = BasicChangeLogDiscriminant<S>;
 
-        match log {
-            BasicChangeLogEntry::Bank(bank_changelog_entry) => {
-                BankMessageGenerator::<S>::assert_state(
-                    config.clone().into(),
-                    bank_changelog_entry,
-                )
-                .await?;
+    async fn assert_state(
+        &self,
+        rollup_state_accessor: Arc<Self::ClientConfig>,
+    ) -> Result<(), anyhow::Error> {
+        match self {
+            BasicChangeLogEntry::Bank(b) => {
+                b.assert_state(Arc::new((*rollup_state_accessor).clone().into()))
+                    .await
             }
-            BasicChangeLogEntry::ValueSetter(value_setter_changelog_entry) => {
-                ValueSetterMessageGenerator::<S>::assert_state(
-                    config.clone().into(),
-                    value_setter_changelog_entry,
-                )
-                .await?;
+            BasicChangeLogEntry::ValueSetter(v) => {
+                v.assert_state(Arc::new((*rollup_state_accessor).clone().into()))
+                    .await
             }
         }
     }
 
-    Ok(())
+    fn as_discriminant(&self) -> Self::Discriminant {
+        match self {
+            BasicChangeLogEntry::Bank(b) => BasicChangeLogDiscriminant::Bank(b.as_discriminant()),
+            BasicChangeLogEntry::ValueSetter(v) => {
+                BasicChangeLogDiscriminant::ValueSetter(v.as_discriminant())
+            }
+        }
+    }
 }
 
 /// The basic configuration for any rollup http client.